requirements.txt
demucs
deep_translator
ffmpeg
openai-whisper
phonemizer
ffmpeg-python
numpy
torch
librosa
deep_translator
torchaudio 
transformers 
python-Levenshtein
